seperate extended battle member and generated member
